molularize commands
