+++ client/src/commands/mod.rs
@@ -1,0 +1,8 @@
+pub mod encointer_bazaar;
+pub mod encointer_ceremonies;
+pub mod encointer_communities;
+pub mod encointer_core;
+pub mod encointer_democracy;
+pub mod encointer_faucet;
+pub mod frame;
+pub mod keystore;
